Add new settings card with tabs
