+++ package.json
@@ -15,6 +15,7 @@
         "start": "node server.js"
     },
     "dependencies": {
+        "@heroicons/react": "^1.0.4",
         "axios": "^0.21.1",
         "express": "^4.17.1",
         "js-cookie": "^2.2.1",
