+++ src/pages/configuration/editor/settings/visual-settings.tsx
@@ -10,6 +10,16 @@ import {
 } from '@components';
 import {icons} from '@assets';
 import {types} from '@types';
+import {
+    CreditCardIcon,
+    OfficeBuildingIcon,
+    UserIcon,
+    UsersIcon,
+} from '@heroicons/react/solid';
+
+function classNames(...classes) {
+    return classes.filter(Boolean).join(' ');
+}
 
 interface Props {
     config: types.SurveyConfig;
@@ -21,122 +31,64 @@ interface Props {
     validation: types.ValidationResult;
 }
 const VisualSettings = (props: Props) => {
-    const [collapse, setCollapse] = useState(true);
-    useEffect(() => setCollapse(true), [props.config.local_id]);
-
-    const [actionLabel, setActionLabel] = useState('');
-
-    const buttonCSS =
-        'w-7 h-7 p-0.5 m-1.5 opacity-70 hover:opacity-100 rounded ringable-dark';
-    const buttons = (
-        <>
-            <button
-                className={buttonCSS}
-                onClick={() => {
-                    props.openDuplicateModal();
-                }}
-                onMouseEnter={() => setActionLabel('duplicate survey')}
-                onFocus={() => setActionLabel('duplicate survey')}
-            >
-                {icons.duplicate}
-            </button>
-            <button
-                className={buttonCSS}
-                onClick={props.openRemoveModal}
-                onMouseEnter={() => setActionLabel('remove survey')}
-                onFocus={() => setActionLabel('remove survey')}
-            >
-                {icons.trash}
-            </button>
-        </>
-    );
+    const tabs = [
+        {name: 'My Account', href: '#', icon: UserIcon, current: false},
+        {name: 'Company', href: '#', icon: OfficeBuildingIcon, current: false},
+        {name: 'Team Members', href: '#', icon: UsersIcon, current: true},
+        {name: 'Billing', href: '#', icon: CreditCardIcon, current: false},
+    ];
 
     return (
-        <EditorFormCard
-            label='General Settings'
-            icon={icons.tune}
-            className='z-20 mt-6'
-            collapse={collapse}
-            setCollapse={setCollapse}
-            buttons={buttons}
-            actionLabel={actionLabel}
-            setActionLabel={setActionLabel}
-            validation={props.validation}
-        >
-            <div className='w-full centering-col gap-y-0.5'>
-                <Label text='Title' />
-                <TextInput
-                    value={props.config.title}
-                    setValue={(newValue: string) => {
-                        props.updateConfig({
-                            ...props.config,
-                            title: newValue,
-                        });
-                    }}
-                    disabled={!props.config.draft || collapse}
-                />
+        <div className='p-4 bg-white rounded shadow flex-col-center'>
+            <div className='sm:hidden'>
+                <label htmlFor='tabs' className='sr-only'>
+                    Select a tab
+                </label>
+                <select
+                    id='tabs'
+                    name='tabs'
+                    className='block w-full border-gray-300 rounded-md focus:ring-indigo-500 focus:border-indigo-500'
+                    defaultValue={tabs.find((tab) => tab.current).name}
+                >
+                    {tabs.map((tab) => (
+                        <option key={tab.name}>{tab.name}</option>
+                    ))}
+                </select>
             </div>
-            <div className='w-full centering-col gap-y-0.5'>
-                <Label text='URL conform identifier' />
-                <TextInput
-                    value={props.config.survey_name}
-                    setValue={(newValue: string) => {
-                        props.updateConfig({
-                            ...props.config,
-                            survey_name: newValue,
-                        });
-                    }}
-                    disabled={!props.config.draft || collapse}
-                />
+            <div className='hidden w-full sm:block'>
+                <div className='w-full border-b border-gray-200'>
+                    <nav
+                        className='-mb-px space-x-8 flex-row-left'
+                        aria-label='Tabs'
+                    >
+                        {tabs.map((tab) => (
+                            <a
+                                key={tab.name}
+                                href={tab.href}
+                                className={classNames(
+                                    tab.current
+                                        ? 'border-blue-500 text-blue-600'
+                                        : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300',
+                                    'group inline-flex items-center py-2 px-1 border-b-2 font-weight-600 text-sm',
+                                )}
+                                aria-current={tab.current ? 'page' : undefined}
+                            >
+                                <tab.icon
+                                    className={classNames(
+                                        tab.current
+                                            ? 'text-blue-500'
+                                            : 'text-gray-400 group-hover:text-gray-500',
+                                        '-ml-0.5 mr-2 h-5 w-5',
+                                    )}
+                                    aria-hidden='true'
+                                />
+                                <span>{tab.name}</span>
+                            </a>
+                        ))}
+                    </nav>
+                </div>
             </div>
-
-            <div className='w-full centering-col gap-y-0.5'>
-                <Label text='Description' />
-                <TextArea
-                    value={props.config.description}
-                    setValue={(newValue: string) => {
-                        props.updateConfig({
-                            ...props.config,
-                            description: newValue,
-                        });
-                    }}
-                    disabled={!props.config.draft || collapse}
-                />
-            </div>
-
-            <div
-                className={'h-0.5 bg-gray-300'}
-                style={{width: 'calc(100% + 1.5rem)'}}
-            />
-
-            <div className='w-full flex-col-left gap-y-0.5'>
-                <Label text='Start survey at' />
-                <DatePicker
-                    timestamp={props.config.start}
-                    setTimestamp={(timestamp: number) => {
-                        props.updateConfig({
-                            ...props.config,
-                            start: timestamp,
-                        });
-                    }}
-                    disabled={!props.config.draft || collapse}
-                />
-            </div>
-
-            <div className='w-full flex-col-left gap-y-0.5'>
-                <Label text='End survey by' />
-                <DatePicker
-                    timestamp={props.config.end}
-                    setTimestamp={(timestamp: number) => {
-                        props.updateConfig({
-                            ...props.config,
-                            end: timestamp,
-                        });
-                    }}
-                    disabled={!props.config.draft || collapse}
-                />
-            </div>
-        </EditorFormCard>
+        </div>
     );
 };
